requirements.txt
# Core ML & embeddings
sentence-transformers>=2.2.2
torch>=2.0.1
torchvision>=0.15.2
torchaudio>=2.0.2

# Text preprocessing & tokenization
nltk>=3.8.1
spacy>=3.5.0
spacy-transformers>=1.2.0

# BM25 ranking
rank-bm25>=0.2.2

# Data handling
pandas>=2.0.3
numpy>=1.25.1
scikit-learn>=1.2.2

# Vector search
faiss-cpu>=1.7.4

# Web scraping (if your scrape_url uses requests + BeautifulSoup)
requests>=2.31.0
beautifulsoup4>=4.12.2
